requirements.txt
pandas>=1.3.0
openpyxl>=3.0.0
xlrd>=2.0.0
psycopg2-binary>=2.9.0
python-dotenv>=0.19.0

